Break user model up; peel off auth, auth, and content associations
